Add support for the SpacePilot Pro
I wasn't sure how to order the button mapping, as the buttons are laid out differently on the SpacePilot Pro compared to others. So I ordered them top to bottom, left and right.

I thought about putting the matching buttons at the same index as on the other devices, but that would lead to weird placement as well, because the "5" button would be separated from "1"-"4" then.